requirements.txt
Flask
mysql-connector-python
flask-cors
python-dotenv
boto3==1.34.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
PyJWT==2.8.0
requests==2.31.0 